+++ src/BrokerLib/dcc/SensorDecoder.cpp
@@ -10,6 +10,9 @@
 
 #include "SensorDecoder.h"
 
+#include <FmtUtils.h>
+#include <JsonUtils.h>
+#include <Log.h>
 #include <Packet.h>
 
 #include "IDevice.h"
@@ -29,17 +32,24 @@ namespace dcclite::broker
 	{
 		m_rclDevice.TryGetINetworkDevice()->Decoder_RegisterPin(*this, m_clPin, "pin");
 
-		auto pullup = params.FindMember("pullup");
-		m_fPullUp = pullup != params.MemberEnd() ? pullup->value.GetBool() : false;
+		m_fPullUp = dcclite::json::TryGetDefaultBool(params, "pullup", false);		
+		m_fInverted = dcclite::json::TryGetDefaultBool(params, "inverted", false);
 
-		auto inverted = params.FindMember("inverted");
-		m_fInverted = inverted != params.MemberEnd() ? inverted->value.GetBool() : false;
+		if (params.HasMember("activateDelay") && params.HasMember("activateDelayMs"))
+		{
+			dcclite::Log::Warn("[SensorDecoder::SensorDecoder] Decoder {} contains both activateDelay and activateDelayMs", this->GetName());
+		}
 
-		auto activateDelay = params.FindMember("activateDelay");
-		m_uActivateDelay = activateDelay != params.MemberEnd() ? activateDelay->value.GetInt() : 0;
+		if (params.HasMember("deactivateDelay") && params.HasMember("deactivateDelayMs"))
+		{
+			dcclite::Log::Warn("[SensorDecoder::SensorDecoder] Decoder {} contains both deactivateDelay and deactivateDelayMs", this->GetName());
+		}
 
-		auto deactivateDelay = params.FindMember("deactivateDelay");
-		m_uDeactivateDelay = deactivateDelay != params.MemberEnd() ? deactivateDelay->value.GetInt() : 0;
+		m_uActivateDelay = dcclite::json::TryGetDefaultInt(params, "activateDelay", 0) * 1000;
+		m_uDeactivateDelay = dcclite::json::TryGetDefaultInt(params, "deactivateDelay", 0) * 1000;
+
+		m_uActivateDelay = dcclite::json::TryGetDefaultInt(params, "activateDelayMs", 0);
+		m_uDeactivateDelay = dcclite::json::TryGetDefaultInt(params, "deactivateDelayMs", 0);
 	}
 
 	SensorDecoder::~SensorDecoder()
@@ -56,8 +66,26 @@ namespace dcclite::broker
 			(m_fPullUp ? dcclite::SensorDecoderFlags::SNRD_PULL_UP : 0) |
 			(m_fInverted ? dcclite::SensorDecoderFlags::SNRD_INVERTED : 0)
 		);
-		packet.Write8(m_uActivateDelay);
-		packet.Write8(m_uDeactivateDelay);
+
+		//Old devices? Send a byte only with a delay in seconds...
+		if (m_rclDevice.TryGetINetworkDevice()->GetProtocolVersion() == dcclite::PROTOCOL_VERSION9)
+		{
+			uint8_t activateDelay = m_uActivateDelay / 1000;
+			uint8_t deactivateDelay = m_uDeactivateDelay / 1000;
+
+			//make sure if activate or deactivate has something (less than 1000), that we have at least 1 sec
+			activateDelay = ((activateDelay == 0) && m_uActivateDelay) ? 1 : activateDelay;			
+			deactivateDelay = ((deactivateDelay == 0) && m_uDeactivateDelay) ? 1 : deactivateDelay;
+				
+			packet.Write8(activateDelay);
+			packet.Write8(deactivateDelay);
+		}
+		else
+		{
+			//new devices handles milliseconds (16 bits) delay
+			packet.Write16(m_uActivateDelay);
+			packet.Write16(m_uDeactivateDelay);
+		}		
 	}
 
 	void SensorDecoder::Serialize(dcclite::JsonOutputStream_t &stream) const