+++ src/Embedded/lib/Shared/Packet.h
@@ -94,7 +94,8 @@ namespace dcclite
 
 	constexpr uint8_t MAX_DECODERS_STATES_PER_PACKET = 64;
 
-	constexpr uint16_t PROTOCOL_VERSION = 9;
+	constexpr uint16_t PROTOCOL_VERSION9 = 9;
+	constexpr uint16_t PROTOCOL_VERSION = 10;
 
 	constexpr uint8_t MAX_NODE_NAME = 16;